Centralize search icons and move JSX components into `.jsx` files
