+++ src/pages/Resources/ResourcesView.jsx
