+++ src/pages/Resources/SearchResult/SearchResult.jsx
