+++ src/pages/dataSets/SearchBox/index.js
@@ -3,8 +3,8 @@ import PropTypes from 'prop-types';
 import { InputGroup, FormControl, Button } from 'react-bootstrap';
 import styled from 'styled-components';
 import xIcon from '../../../assets/img/xmark-solid.svg';
-import SearchIconImg from './search.svg';
-import CloseIconImg from './close.svg';
+import SearchIconImg from '../../../assets/icons/search.svg';
+import CloseIconImg from '../../../assets/icons/close.svg';
 
 const Container = styled.div`
   width: 100%;
